Extract EjbcaConnectionFactory for the IT-only EJBCA connection/SSL code and coverage-exclude it
